sdhci: Remove Samsung's broken hack
To avoid a divide-by-zero, Samsung returns a 0 if
host->clock is not set.

This is bad - the timeout will (of course) now be
as short as possible.

Mainline does it differently - It uses a base
timeout value regardless of whether host->clock
exists, then adds that value if it exists.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>

thx to Entropy512